requirements.txt
google-play-scraper
langdetect
openai
google-generativeai
jinja2
google-api-python-client
google-auth-oauthlib
python-dotenv
streamlit
pandas
schedule
